requirements.txt
openai
supabase
python-dotenv
openai-agents
psycopg2-binary
requests
